the project is done that only a few days

+++ restauran.console/restauran.console/Interfaces/Services/IOrderService.cs
@@ -5,6 +5,6 @@ namespace restauran.console.Interfaces.Services
     public interface IOrderService
     {
         public Task<IEnumerable<OrderViewModel>> GetAllAsync();
-        public Task<OrderViewModel> GetAsync(long id);
+        public Task<OrderViewModel> GetAsync(int id);
     }
 }
